step7 testCode: 클래스별로 파일을 만듬 ArrayParser 클래스의 메인 메소드가 잘 동작하는지 테스트 코드 작성
